Added work page
